requirements.txt
# ---- Core framework ----
fastapi
uvicorn

# ---- Settings and Utilities ----
os
python-dotenv
pydantic
pydantic-core 

# ---- Database + ORM ----
sqlalchemy
asyncpg

# ---- Security & Auth ----
pyjwt
python-jose
passlib
argon2-cffi

# --- Other modules ----
datetime
typing
httpx
pytest
asyncio